+++ Makefile
@@ -124,9 +124,11 @@ copy-public-builds:
 
 
 .PHONY: generate
-generate:
-	rm -rf packages/db/pkg/queries/*
-	sqlc generate
+generate: generate/api generate/orchestrator generate/template-manager generate/envd generate/db
+generate/%:
+	@echo "Generating code for *$(notdir $@)*"
+	$(MAKE) -C packages/$(notdir $@) generate
+	@printf "\n\n"
 
 .PHONY: migrate
 migrate:
